connecting asteroid details screen to app

+++ app/src/main/java/com/jar/jasteroidradar/ui/navigation/Screen.kt
@@ -14,7 +14,7 @@ sealed class Screen(val route: String = ""){
         }
     }
     object AsteroidDetails: Screen(route = "$ASTEROID_DETAILED_KEY/{asteroidId}") {
-        fun asteroidId(asteroidId: Int): String {
+        fun asteroidId(asteroidId: Long): String {
             return "$ASTEROID_DETAILED_KEY/$asteroidId"
         }
     }
